refactor(cli): reorganize utilities - moved utilities from the command folder to the common directory - separated utilities based on purpose
