+++ packages/cli/src/common/resolve/index.ts
@@ -1,6 +1,7 @@
 import { type ConfigLoaderSuccessResult, createMatchPath } from "tsconfig-paths"
 
-import { ErrorMap } from "../../common/error"
+import { ErrorMap } from "@/common/error"
+import type { ConfigType } from "@/common/types"
 export async function resolveImport(
   importPath: string,
   config: Pick<ConfigLoaderSuccessResult, "absoluteBaseUrl" | "paths">,
@@ -19,4 +20,16 @@ export async function resolveImport(
     })
   }
   return match
+}
+
+export async function resolveAllPaths(
+  config: ConfigType,
+  tsconfig: ConfigLoaderSuccessResult,
+) {
+  return {
+    utils: await resolveImport(config.utils, tsconfig),
+    components: await resolveImport(config.components, tsconfig),
+    hooks: await resolveImport(config.hooks, tsconfig),
+    styledsystem: await resolveImport(config.styledsystem, tsconfig),
+  }
 }
