--- packages/cli/src/init/utils/checkJsonInit.ts
@@ -1,6 +1,0 @@
-import fs from "fs-extra"
-import path from "path"
-
-export async function checkJsonInit(root: string) {
-  return await fs.pathExists(path.join(root, "components.json"))
-}
