--- packages/cli/src/init/utils/initTemplate.ts
@@ -1,13 +1,0 @@
-import fs from "fs-extra"
-import path from "path"
-
-import { template } from "./template"
-
-export function initTemplate(dir: string) {
-  try {
-    const tokenPath = path.join(dir, "tokens.ts")
-    fs.writeFileSync(tokenPath, template)
-  } catch (e) {
-    throw new Error(`Failed to create tokens.ts file: ${e}`)
-  }
-}
